requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Dance==7.0.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
SQLAlchemy==1.4.53
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
requests==2.31.0
oauthlib==3.2.2
requests-oauthlib==1.3.1
cryptography==41.0.7
reportlab==4.0.7
xhtml2pdf==0.2.16
